gcp/requirements.txt
moviepy==1.0.1
pydub==0.23.1
Wave==0.0.2
google-cloud-speech==1.3.2 
google-cloud-storage==1.26.0 